requirements.txt
asgiref==3.2.10
djangorestframework==3.11.1
django==2.2.0
django-heroku==0.3.1
gunicorn==20.0.4
pip==19.0.3
psycopg2==2.8.4
pytz==2020.1
setuptools==40.8.0
sqlparse==0.3.1